\name{loeweData}
\alias{loeweData}
\docType{data}
\title{loeweData Dataset}
\description{
  Dataset obtained as a result of using function \code{doNls}, based on data 
  in the text file \dQuote{mixlowExampleData.txt}.  This file contains made-up
  example data. 
}
\usage{data(loeweData)}
\format{
See help file for function \code{doLoewe} for composition of loeweData
dataset.
}
\details{
  The text file \dQuote{mixlowExampleData.txt} provides example data for 
  a cytotoxicity experiment in which two drugs and their mixture is 
  tested.  Each drug is tested in three replicate trays.
  }


\seealso{\code{\link{doNls}}}
            

\examples{
data(loeweData)
}
\keyword{datasets}
