\name{mixlowData}
\alias{mixlowData}
\docType{data}
\title{mixlowData Dataset}
\description{
  Dataset obtained as a result of using function \code{prepareData}, based on data 
  in the text file \dQuote{mixlowExampleData.txt}.  This file contains made-up
  example data. 
}
\usage{data(mixlowData)}
\format{
See help file for function \code{prepareData} for composition of mixlowData
dataset.

}
\details{
  The text file \dQuote{mixlowExampleData.txt} provides example data for 
  a cytotoxicity experiment in which two drugs and their mixture is 
  tested.  Each drug is tested in three replicate trays.
  }


\seealso{\code{\link{prepareData}}}
           

\examples{
data(mixlowData)
}
\keyword{datasets}
