\name{nlmeData}
\alias{nlmeData}
\docType{data}
\title{nlmeData Dataset}
\description{
  Dataset obtained as a result of using function \code{doNlme}, based on data 
  in the text file \dQuote{mixlowExampleData.txt}.  This file contains made-up
  example data. 
}
\usage{data(nlmeData)}
\format{
See help file for function \code{doNlme} for composition of nlmeData
dataset.
}
\details{
  The text file \dQuote{mixlowExampleData.txt} provides example data for 
  a cytotoxicity experiment in which two drugs and their mixture is 
  tested.  Each drug is tested in three replicate trays.
  }



\examples{
data(nlmeData)
}
\keyword{datasets}
