\name{trayData}
\alias{trayData}
\docType{data}
\title{trayData Dataset}
\description{
  Dataset obtained as a result of using function \code{readDataFile}, based on data 
  in the text file \dQuote{mixlowExampleData.txt}.  This file contains made-up
  example data. 
}
\usage{data(trayData)}
\format{
See help file for function \code{readDataFile} for composition of trayData
dataset.
}
\details{
  The text file \dQuote{mixlowExampleData.txt} provides example data for 
  a cytotoxicity experiment in which two drugs and their mixture is 
  tested.  Each drug is tested in three replicate trays.
  }


\seealso{\code{\link{readDataFile}}}
            

\examples{
data(trayData)
}
\keyword{datasets}
