\name{nlsData}
\alias{nlsData}
\docType{data}
\title{nlsData Dataset}
\description{
  Dataset obtained as a result of using function \code{doNls}, based on data 
  in the text file \dQuote{mixlowExampleData.txt}.  This file contains made-up
  example data. 
}
\usage{data(nlsData)}
\format{
See help file for function \code{doNls} for composition of nlsData
dataset.
}
\details{
  The text file \dQuote{mixlowExampleData.txt} provides example data for 
  a cytotoxicity experiment in which two drugs and their mixture is 
  tested.  Each drug is tested in three replicate trays.
  }


\seealso{\code{\link{doNls}}}
            

\examples{
data(nlsData)
}
\keyword{datasets}
